+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Include match association records in facility history list [#851](https://github.com/open-apparel-registry/open-apparel-registry/pull/851)
+- Include facility claim data in facility history list [#852](https://github.com/open-apparel-registry/open-apparel-registry/pull/852)
 
 ### Deprecated